analysis/pyproject.toml
[project]
name = "seras-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "ipykernel>=7.1.0",
    "japanize-matplotlib>=1.1.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
]
